Finish up lab

+++ lab2/lab5.h
@@ -6,8 +6,8 @@ extern const int MAX;
 int readTotal();
 void fillArray(int total, int *array);
 void selectionSort(int *array, int total);
-int calcMean(int total, int *array);
-int calcMedian(int *array, int total);
-void displayResults(int mean, int median);
+double calcMean(int total, int *array);
+double calcMedian(int *array, int total);
+void displayResults(double mean, double median);
 
 #endif
